Move custom formatting docs to Logger.Formatter
