+++ lib/logger/lib/logger/formatter.ex
@@ -4,12 +4,21 @@ defmodule Logger.Formatter do
   @moduledoc ~S"""
   Conveniences for formatting data for logs.
 
-  This module allows developers to specify a string that
-  serves as template for log messages, for example:
+  This module allows developers to specify a `{module, function}`
+  or a string that serves as template for log messages.
 
-      $time $metadata[$level] $message\n
+  ## Formatting string
 
-  Will print error messages as:
+  The log messages can be controlled by a formatting string.
+  Here is an example of how to configure the `:console` backend
+  in a `config/config.exs` file:
+
+  For example:
+
+      config :logger, :console,
+        format: "$time $metadata[$level] $message\n"
+
+  The above will print error messages as:
 
       18:43:12.439 user_id=13 [error] Hello\n
 
@@ -27,6 +36,47 @@ defmodule Logger.Formatter do
   which compiles the string into a format for fast operations at
   runtime and `format/5` to format the compiled pattern into an
   actual IO data.
+
+  ## Formatting function
+
+  You can also customize the format of your log messages to a
+  `{module, function}` tuple if you wish to provide your own
+  format function. Here is an example of how to configure the
+  `:console` backend in a `config/config.exs` file:
+
+      config :logger, :console,
+        format: {MyConsoleLogger, :format}
+
+  And here is an example of how you can define `MyConsoleLogger.format/4`
+  from the above configuration:
+
+      defmodule MyConsoleLogger do
+        def format(level, message, timestamp, metadata) do
+          # Custom formatting logic...
+        end
+      end
+
+  It is extremely important that **the formatting function does
+  not fail**, as it will bring that particular logger instance down,
+  causing your system to temporarily lose messages. If necessary,
+  wrap the function in a `rescue` and log a default message instead:
+
+      defmodule MyConsoleLogger do
+        def format(level, message, timestamp, metadata) do
+          # Custom formatting logic...
+        rescue
+          _ -> "could not format: #{inspect({level, message, metadata})}"
+        end
+      end
+
+  The `{module, function}` will be invoked with four arguments:
+
+    * the log level: an atom
+    * the message: this is usually chardata, but in some cases it
+      may contain invalid data. Since the formatting function should
+      *never* fail, you need to prepare for the message being anything
+    * the current timestamp: a term of type `t:time/0`
+    * the metadata: a keyword list
 
   ## Metadata
